requirements.txt
tgcrypto
flask
pyrogram
asyncio
python-dateutil